backend-api/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
django-environ>=0.4.5
django-extensions
django-cors-headers
djangorestframework>=3.12.1
remote_pdb>=2.0.0
